# docs/swagger.yaml
basePath: /
definitions:
...
  handler.FPReportSyncRequest:
    properties:
      facility_ids:
        items:
          type: string
        type: array
      force:
        type: boolean
      period:
        type: string
    type: object
...
  service.CreateRegistrationInput:
    properties:
      age:
        type: integer
      breast_cancer_screening:
        type: string
      cervical_cancer_status:
        type: string
      cervical_cancer_treatment:
        type: string
      cervical_screening_method:
        type: string
      condoms_female_units:
        type: integer
      condoms_male_units:
        type: integer
      counseling_as_couple:
        type: boolean
      counseling_individual:
        type: boolean
      counseling_ms:
        type: boolean
      counseling_om:
        type: boolean
      counseling_se:
        type: boolean
      counseling_wd:
        type: boolean
      district:
        type: string
      fam_lam:
        type: boolean
      fam_standard_days:
        type: boolean
      fam_two_day:
        type: boolean
      given_name:
        type: string
      hts_code:
        type: string
      implant_3_years:
        type: boolean
      implant_5_years:
        type: boolean
      implant_removal_reason:
        type: string
      implant_removal_timing:
        type: string
      injectable_dmpa_im_doses:
        type: integer
      injectable_dmpa_sc_pa_doses:
        type: integer
      injectable_dmpa_sc_si_doses:
        type: integer
      is_new_user:
        type: boolean
      is_revisit:
        type: boolean
      is_switching:
        type: boolean
      is_visitor:
        type: boolean
      iud_copper_t:
        type: boolean
      iud_hormonal_3_years:
        type: boolean
      iud_hormonal_5_years:
        type: boolean
      iud_removal_reason:
        type: string
      iud_removal_timing:
        type: string
      nin:
        type: string
      parish:
        type: string
      phone_number:
        type: string
      pills_coc_cycles:
        type: integer
      pills_ecp_pieces:
        type: integer
      pills_pop_cycles:
        type: integer
      post_abortion_fp_timing:
        type: string
      postpartum_fp_timing:
        type: string
      previous_method:
        type: string
      referral_number:
        type: string
      referral_reason:
        type: string
      remarks:
        type: string
      screened_for_sti:
        type: boolean
      sex:
        type: string
      side_effects:
        type: string
      subcounty:
        type: string
      surname:
        type: string
      switching_reason:
        type: string
      tubal_ligation:
        type: boolean
      vasectomy:
        type: boolean
      village:
        type: string
      visit_date:
        type: string
    type: object
...
  utils.APIError:
    properties:
      errors:
        items:
          $ref: '#/definitions/utils.ErrorDetail'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
  utils.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  utils.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
info:
  contact: {}
  description: Uganda MoH HMIS MCH 007 Integrated Family Planning Register
  title: HMIS MCH 007 – Integrated FP Register API
  version: "1.0"
paths:
...
  /api/v1/registrations:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      - description: Filter by visit date (YYYY-MM-DD)
        in: query
        name: visit_date
        type: string
      - description: Search by name, NIN, or client number
        in: query
        name: search
        type: string
      - description: Filter by sex (M/F)
        in: query
        name: sex
        type: string
      - description: From date
        in: query
        name: date_from
        type: string
      - description: To date
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List FP register entries
      tags:
      - registrations
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateRegistrationInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - BearerAuth: []
      summary: Create a new FP register entry
      tags:
      - registrations
  /api/v1/registrations/{id}:
    delete:
      parameters:
      - description: Registration ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Soft-delete an FP register entry
      tags:
      - registrations
    get:
      parameters:
      - description: Registration ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get a single FP register entry
      tags:
      - registrations
    put:
      consumes:
      - application/json
      parameters:
      - description: Registration ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateRegistrationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update an FP register entry
      tags:
      - registrations
  /api/v1/reports/family-planning/monthly:
    get:
      parameters:
      - description: Period in YYYYMM format
        in: query
        name: period
        required: true
        type: string
      - description: Facility ID
        in: query
        name: facility_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Monthly FP methods report
      tags:
      - reports
  /api/v1/reports/family-planning/payload-preview:
    get:
      parameters:
      - description: Period in YYYYMM format
        in: query
        name: period
        required: true
        type: string
      - description: Facility ID
        in: query
        name: facility_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Preview DHIS2 payload for monthly FP report
      tags:
      - reports
  /api/v1/reports/family-planning/sync:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sync payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.FPReportSyncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Sync monthly FP report to DHIS2
      tags:
      - reports
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"